refactor: rename module agent_reach → hivereach
Move package directory and update all 100 references across 26 files
(imports, pyproject build config, docs, tests). No behavior change.
