+++ CLAUDE.md
@@ -10,19 +10,19 @@ Repo: github.com/Panniantong/Agent-Reach | License: MIT | Version: 1.3.0
 - `pytest tests/ -v` — All tests
 - `pytest tests/test_cli.py -v` — CLI tests only
 - `bash test.sh` — Full integration test (creates venv, installs, runs doctor + channel tests)
-- `python -m agent_reach.cli doctor` — Run diagnostics
-- `python -m agent_reach.cli install --env=auto` — Auto-configure
+- `python -m hivereach.cli doctor` — Run diagnostics
+- `python -m hivereach.cli install --env=auto` — Auto-configure
 
 ## Structure
-- `agent_reach/cli.py` — CLI entry point (argparse)
-- `agent_reach/core.py` — Core read/search routing logic
-- `agent_reach/config.py` — Config management (YAML, env vars)
-- `agent_reach/doctor.py` — Diagnostics engine
-- `agent_reach/channels/` — One file per platform (twitter.py, reddit.py, youtube.py, etc.)
-- `agent_reach/channels/base.py` — Base channel class (all channels inherit from this)
-- `agent_reach/integrations/mcp_server.py` — MCP server integration
-- `agent_reach/skill/` — OpenClaw skill files
-- `agent_reach/guides/` — Usage guides
+- `hivereach/cli.py` — CLI entry point (argparse)
+- `hivereach/core.py` — Core read/search routing logic
+- `hivereach/config.py` — Config management (YAML, env vars)
+- `hivereach/doctor.py` — Diagnostics engine
+- `hivereach/channels/` — One file per platform (twitter.py, reddit.py, youtube.py, etc.)
+- `hivereach/channels/base.py` — Base channel class (all channels inherit from this)
+- `hivereach/integrations/mcp_server.py` — MCP server integration
+- `hivereach/skill/` — OpenClaw skill files
+- `hivereach/guides/` — Usage guides
 - `tests/` — pytest tests
 - `config/mcporter.json` — MCP tool config
