+++ hivereach/doctor.py
@@ -5,8 +5,8 @@ Each channel knows how to check itself. Doctor just collects the results.
 """
 
 from typing import Dict
-from agent_reach.config import Config
-from agent_reach.channels import get_all_channels
+from hivereach.config import Config
+from hivereach.channels import get_all_channels
 
 
 def check_all(config: Config) -> Dict[str, dict]:
