+++ hivereach/scripts/transcribe_xiaoyuzhou.sh
