+++ tests/test_twitter_channel.py
@@ -2,7 +2,7 @@
 
 from unittest.mock import patch, Mock
 
-from agent_reach.channels.twitter import TwitterChannel
+from hivereach.channels.twitter import TwitterChannel
 
 
 def _cp(stdout="", stderr="", returncode=0):
