+++ tests/test_xiaoyuzhou_install.py
@@ -2,7 +2,7 @@
 
 from unittest.mock import patch
 
-import agent_reach.cli as cli
+import hivereach.cli as cli
 
 
 class _DummyConfig:
@@ -11,7 +11,7 @@ class _DummyConfig:
 
 
 def test_install_xiaoyuzhou_deps_does_not_raise_when_no_groq_key(capsys):
-    with patch("agent_reach.config.Config", return_value=_DummyConfig()), \
+    with patch("hivereach.config.Config", return_value=_DummyConfig()), \
          patch("os.path.isfile", side_effect=lambda p: True if str(p).endswith("transcribe.sh") else False), \
          patch("shutil.which", return_value=None):
         cli._install_xiaoyuzhou_deps()
